requirements-dev.txt
-r requirements.txt

packaging==16.8
py==1.4.33
pyparsing==2.2.0
pytest==3.0.7
pytest-django==3.1.2
pytest-mock==1.6.0
mock==2.0.0
six==1.10.0
bumpversion==0.5.3
